// admin/src/api/generated/schemas/adminCurrencyItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */
import type { AdminCurrencyFeatureSettings } from './adminCurrencyFeatureSettings';

export type AdminCurrencyItem = {
  asset_code?: string;
  asset_name?: string;
  status?: number;
  settings?: AdminCurrencyFeatureSettings;
  chain_total?: number;
  chain_enabled?: number;
  updated_at?: string;
  icon_url?: string;
  precision?: number;
};